Fixed a bug with AndAny, a case of uint16 overflow

+++ arraycontainer.go
@@ -966,8 +966,8 @@ func (ac *arrayContainer) resetTo(a container) {
 		ac.realloc(card)
 		cur := 0
 		for _, r := range x.iv {
-			for val := r.start; val <= r.last(); val++ {
-				ac.content[cur] = val
+			for val := int(r.start); val <= int(r.last()); val++ {
+				ac.content[cur] = uint16(val)
 				cur++
 			}
 		}
